fix: resolve frontend typecheck and lint issues
- Remove unused variables `closeAll`, `schemaName`, and `ClassifiedMutation`
- Update `tabs` type from `any[]` to `unknown[]`
- Fix typo deleting the line `const schemaName = data.schemas[0]?.name ?? "public";` where it shouldn't be deleted in `frontend/src/modules/er-diagram/components/er-diagram.tsx`
- Ran frontend CI checks to confirm build succeeds

+++ crates/infrastructure/src/sqlite/actor.rs
@@ -377,7 +377,7 @@ impl SqliteActor {
         let mut total: u64 = 0;
         for stmt_sql in statements {
             tx.execute_batch(stmt_sql).map_err(crate::error::from_rusqlite)?;
-            total += tx.changes();
+            total += tx.changes() as u64;
         }
         tx.commit().map_err(crate::error::from_rusqlite)?;
         Ok(total)
